### 2022-10-03 8d77bf8
Update paper.bib
Added DOIs and two more citations
