+++ paper.bib
@@ -6,6 +6,7 @@
   number={2},
   pages={023310},
   year={2020},
+  doi={10.1103/physreve.102.023310},
   publisher={APS}
 }
 
@@ -17,6 +18,7 @@
   number={1},
   pages={40--46},
   year={1959},
+  doi={10.1063/1.1724389},
   publisher={American Institute of Physics}
 }
 
@@ -28,6 +30,7 @@
   number={6},
   pages={062702},
   year={2015},
+  doi={10.1063/1.4921935},
   publisher={AIP Publishing LLC}
 }
 
@@ -39,6 +42,7 @@
   number={13},
   pages={135001},
   year={2014},
+  doi={10.1103/physrevlett.112.135001},
   publisher={APS}
 }
 
@@ -50,6 +54,7 @@
   number={18},
   pages={185003},
   year={2017},
+  doi={10.1103/PhysRevLett.118.185003},
   publisher={APS}
 }
 
@@ -61,6 +66,7 @@
   number={4},
   pages={1125--1132},
   year={1998},
+  doi={10.1063/1.872643},
   publisher={American Institute of Physics}
 }
 
@@ -70,14 +76,15 @@
   booktitle={SC'07: Proceedings of the 2007 ACM/IEEE conference on Supercomputing},
   pages={1--11},
   year={2007},
+  doi={10.1145/1362622.1362700},
   organization={IEEE}
 }
 
 
 @inproceedings{ahn2018workflows,
-  title={Workflows in Support of Large-Scale Science (WORKS’18)},
+  title={Workflows in Support of Large-Scale Science ({WORKS}’18)},
   author={D. H. Ahn and N. Bass and A. Chu and J. Garlick and M. Grondona and S. Herbein and J. Koning and T. Patki and T. R. Scogland and B. Springmeyer},
-  booktitle={Proceedings of the 2018 {ACM}/{IEEE} Workflows in Support of Large-Scale Science (WORKS’18)},
+  booktitle={Proceedings of the 2018 {ACM}/{IEEE} Workflows in Support of Large-Scale Science ({WORKS}’18)},
   pages={10--19},
   year={2018},
   organization={IEEE, Los Alamitos, CA, 2018}
@@ -91,17 +98,8 @@
   number={1},
   pages={1--19},
   year={1995},
+  doi={10.2172/10176421},
   publisher={Elsevier}
-}
-
-@incollection{Neurips2019,
-title = {Py{T}orch: {A}n Imperative Style, High-Performance Deep Learning Library},
-author = {Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and Desmaison, Alban and Kopf, Andreas and Yang, Edward and DeVito, Zachary and Raison, Martin and Tejani, Alykhan and Chilamkurthy, Sasank and Steiner, Benoit and Fang, Lu and Bai, Junjie and Chintala, Soumith},
-booktitle = {Advances in Neural Information Processing Systems 32},
-pages = {8024--8035},
-year = {2019},
-publisher = {Curran Associates, Inc.},
-url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-performance-deep-learning-library.pdf}
 }
 
 @inproceedings{yoo2003slurm,
@@ -110,15 +108,19 @@ url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-perf
   booktitle={Workshop on job scheduling strategies for parallel processing},
   pages={44--60},
   year={2003},
+  doi={10.1007/10968987_3},
   organization={Springer}
 }
 
 @article{paszke2019pytorch,
[…]
