### 2022-10-03 41f80c7
Update paper.md
Added a few more citations
