### 2022-10-04 #4782 @danielskatz
👋 @apachalieva - note that the reference check looks at all references in your .bib file. If you are not using some of them in your paper, you can remove them from this file, then not worry about if they have DOIs. On the other hand, if the references associated with the invalid DOIs are needed in your paper, you will need to fix the invalid DOIs.
