### 2022-10-04 b8ff3db
Update paper.bib
Invalid DOIs updated
