+++ paper.bib
@@ -119,7 +119,7 @@
   pages = {8024--8035},
   year = {2019},
   publisher = {Curran Associates, Inc.},
-  doi={10.5555/3454287.3455008},
+  doi={https://dl.acm.org/doi/10.5555/3454287.3455008}
   url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-performance-deep-learning-library.pdf}
 }
 
@@ -130,7 +130,7 @@
   volume={12},
   pages={2825--2830},
   year={2011},
-  doi={10.5555/1953048.2078195},
+  doi={https://dl.acm.org/doi/10.5555/1953048.2078195},
   publisher={JMLR. org}
 }
 
@@ -199,7 +199,7 @@ author={{Los Alamos National Laboratory}}}
   number={1},
   pages={23--31},
   year={1987},
-  doi={10.5555/324493.325074},
+  doi={https://dl.acm.org/doi/pdf/10.5555/324493.325074},
   publisher={IEEE}
 }
 
