### 2022-10-04 10bd3e4
Update paper.bib
Invalid DOIs deleted
