### 2022-10-06 #4822 @editorialbot
[REVIEW]: GLUE Code: A framework handling communication and interfaces between scales
**Submitting author:** @apachalieva (<a href="http://orcid.org/0000-0003-1246-0410">Aleksandra Pachalieva</a>)
**Repository:** https://github.com/lanl/GLUE
**Branch with paper.md** (empty if default branch): manuscript
**Version:** 1.1
