### 2022-10-20 #4822 @danielskatz
👋 @govarguz and @keipertk - I'm just checking in after 2 weeks. Please do get started when you, by generating your checklists. To do this, use the command @editorialbot generate my checklist to create your review checklist. @editorialbot commands need to be the first thing in a new comment.
