### 2022-10-25 lanl/GLUE#2 @keipertk
Contribution Guidelines for JOSS
Hello,
Very nice work here!

I'm looking for community guideline documentation, the last item on my review at openjournals/joss-reviews#4822

Here is a description:
https://joss.readthedocs.io/en/latest/review_criteria.html?highlight=guidelines#community-guidelines

In other projects I've often seen the guidelines listed in a top-level CONTRIBUTING.md file.
