+++ README.md
@@ -24,3 +24,7 @@ The GLUE Code is provided under a BSD-3 license. See [LICENSE](https://github.co
 © 2021. Triad National Security, LLC. All rights reserved.
 
 This program was produced under U.S. Government contract 89233218CNA000001 for Los Alamos National Laboratory (LANL), which is operated by Triad National Security, LLC for the U.S. Department of Energy/National Nuclear Security Administration. All rights in the program are reserved by Triad National Security, LLC, and the U.S. Department of Energy/National Nuclear Security Administration. The Government is granted for itself and others acting on its behalf a nonexclusive, paid-up, irrevocable worldwide license in this material to reproduce, prepare derivative works, distribute copies to the public, perform publicly and display publicly, and to permit others to do so.
+
+## Development and questions
+
+If you would like to contribute to DPFEHM, please clone the repo and make a pull request. If you have any questions, please contact Aleksandra Pachalieva (apachalieva@lanl.gov) or Robert Pavel (rspavel@lanl.gov).
