### 2022-10-25 lanl/GLUE#2 @apachalieva
Hi @keipertk,

I added a "Development and questions" section in the README.md-file in the GLUE Code repository. Let me know if that is sufficient as community guideline documentation.

Thank you!
