### 2022-10-25 84ac185
Update README.md
Small change to the development and questions section
