### 2022-10-25 lanl/GLUE#2 @keipertk
Thanks for pointing that out, for whatever reason I didn't scroll past the license section. LGTM
