### 2022-11-10 #4822 @danielskatz
👋 @govarguz - it's now been ~5 weeks. Please do get started when you can by generating your checklist. To do this, use the command @editorialbot generate my checklist to create your review checklist. @editorialbot commands need to be the first thing in a new comment.

If you have any issues or concerns about the process, please let me know.
