### 2022-11-12 lanl/GLUE#7 @govarguz
[Documentation] GLUE
I tried to compile the documentation, but I cannot find it. Could you have a look?

Part of https://github.com/openjournals/joss-reviews/issues/4822
