### 2022-11-12 lanl/GLUE#8 @govarguz
[minor] The required packages are not listed
As the matter of reference, I could not find the required packages for using GLUE within the documentation, nor README. This is essential for the users.

Part of https://github.com/openjournals/joss-reviews/issues/4822
