### 2022-11-12 lanl/GLUE#9 @govarguz
[example] add another example, like a tutorial,  jupyter notebook would be an option for this
Hi again, I saw the example and I got an idea on how it works, however, there are plenty functions e.g. in alInterface.py and the rest of the GLUcode which are not pretty clear to me (this is also related to the documentation part).
Adding a step-by-step example on how to use it in LAMMPS would improve also the documentation. Here I would suggest to use a jupyter notebook, or similar to better visualize the workflow.

Part of https://github.com/openjournals/joss-reviews/issues/4822
