### 2022-11-12 lanl/GLUE#10 @govarguz
[title] could be missleading
Regarding the title, "GLUE Code: A framework handling communication and interfaces between scales"
Here the word "interface" sounds a bit broad, between multiscale simulations there are several types of interfaces, even real physical ones, hence I would suggest to write API explicitly.

Part of https://github.com/openjournals/joss-reviews/issues/4822
