### 2022-11-12 lanl/GLUE#11 @govarguz
[training] add more description on how to generate training and the data contained in the repo
Regarding the training part, the description of it, and how was obtained, or how to obtain it, could be added in a couple of lines to enhance readability of the Repo

Part of https://github.com/openjournals/joss-reviews/issues/4822
