### 2022-11-12 #4822 @govarguz
Dear @apachalieva,
there is a new #PR with the general review of GLUE, and a couple of issues I raised also at GLUE's repo.
Best,
Horacio.
cc: @danielskatz
