### 2022-11-14 45ddfa9
Added Doxygen Logic to CMake
Added doxygen logic to CMake and cleaned up messy whitespace
