+++ GLUECode_Library/src/include/alInterface.h
@@ -163,198 +163,198 @@ typedef void * dbHandle_t;
 extern "C"
 {
 #endif
-    /**
-     * @brief C++ function that generates a single fine grain simulation request for ICF using BGK format
-     *
-     * @param input Array of request
-     * @param mpiRank MPI Rank of requesting process
-     * @param tag Tag corresponding to set of requests
-     * @param dbHandle Database to write to
-     * @return bgk_result_t single fine grain simulation result for ICF using BGK format
-     */
+	/**
+	 * @brief C++ function that generates a single fine grain simulation request for ICF using BGK format
+	 *
+	 * @param input Array of request
+	 * @param mpiRank MPI Rank of requesting process
+	 * @param tag Tag corresponding to set of requests
+	 * @param dbHandle Database to write to
+	 * @return bgk_result_t single fine grain simulation result for ICF using BGK format
+	 */
 	bgk_result_t bgk_req_single(bgk_request_t input, int mpiRank, char * tag, dbHandle_t dbHandle);
 
 	/**
-     * @brief C++ function that generates a single fine grain simulation request for ICF using BGK format
-     * with specified request type
-     *
-     * @param input Array of request
-     * @param mpiRank MPI Rank of requesting process
-     * @param tag Tag corresponding to set of requests
-     * @param dbHandle Database to write to
-     * @param reqType ALInterfaceMode_e of request type
-     * @return bgk_result_t single fine grain simulation result for ICF using BGK format
-     */
+	 * @brief C++ function that generates a single fine grain simulation request for ICF using BGK format
+	 * with specified request type
+	 *
+	 * @param input Array of request
+	 * @param mpiRank MPI Rank of requesting process
+	 * @param tag Tag corresponding to set of requests
+	 * @param dbHandle Database to write to
+	 * @param reqType ALInterfaceMode_e of request type
+	 * @return bgk_result_t single fine grain simulation result for ICF using BGK format
+	 */
 	bgk_result_t bgk_req_single_with_reqtype(bgk_request_t input, int mpiRank, char * tag, dbHandle_t dbHandle, unsigned int reqType);
 
-    /**
-     * @brief C++ function that generates a batch of fine grain simulation requests for ICF using BGK format
-     *
-     * @param input Array of request of length numInputs
-     * @param numInputs Length of input array
-     * @param mpiRank MPI Rank of requesting process
-     * @param tag Tag corresponding to set of requests
-     * @param dbHandle Database to write to
-     * @return bgk_result_t array of length numInputs with fine grain simulation results for ICF using BGK format
-     */
+	/**
+	 * @brief C++ function that generates a batch of fine grain simulation requests for ICF using BGK format
+	 *
+	 * @param input Array of request of length numInputs
+	 * @param numInputs Length of input array
+	 * @param mpiRank MPI Rank of requesting process
+	 * @param tag Tag corresponding to set of requests
+	 * @param dbHandle Database to write to
+	 * @return bgk_result_t array of length numInputs with fine grain simulation results for ICF using BGK format
+	 */
 	bgk_result_t* bgk_req_batch(bgk_request_t *input, int numInputs, int mpiRank, char * tag, dbHandle_t dbHandle);
 
 	/**
-     * @brief C++ function that generates a batch of fine grain simulation requests for ICF using BGK format
-     * with specified request type
-     *
-     * @param input Array of request of length numInputs
-     * @param numInputs Length of input array
-     * @param mpiRank MPI Rank of requesting process
-     * @param tag Tag corresponding to set of requests
-     * @param dbHandle Database to write to
-     * @param reqType ALInterfaceMode_e of request type
[…]
