### 2022-11-14 lanl/GLUE#10 @rspavel
Is this a blocking issue regarding the review? We've used this nomenclature throughout many reports as well as publications (preprint and accepted) and would rather stay consistent. And while I agree that "interface" is an incredibly overloaded term, so is "scale"* and it is really the context that indicates which definition is being used.

*: actually, in this context that simultaneously applies to the granularity of the scientific simulation and the number of nodes the simulation is running across on the cluster. Let's pretend that was being clever...
