### 2022-11-14 lanl/GLUE#10 @govarguz
This is am open issue, as you see, I suggested API, but if you explain it better than with a very ambiguous word "interface"...go ahead!
