### 2022-11-17 lanl/GLUE#9 @rspavel
We looked into Jupyter (as well as a potential installation of the service as a "normal" python package) but it did not seem to be a good fit due to the use of the python based `GLUECode_Service` as, well, a service. But I did improve the example scripts and add one to use the training data generation script.
