### 2022-11-22 lanl/GLUE#8 @rspavel
Is there an example of a good practice of formatting this?

On the `GLUECode_Service` side we use a pretty bog standard `requirements.txt` https://github.com/lanl/GLUE/blob/main/GLUECode_Service/requirements.txt

And on the `GLUECode_Library` we list the dependencies both in the https://github.com/lanl/GLUE/blob/main/GLUECode_Library/README.md as well as being part of the CMake build system itself.
