### 2022-11-22 #4822 @apachalieva
Hi @danielskatz, my colleagues and I are working on the open issues. We have divided the tasks among our team and requested additional information from the reviewer, wherever necessary. The open issues do not include any implementation work; therefore, we hope to resolve them by the end of next week.
