### 2022-11-22 lanl/GLUE#14 @rspavel
[Manuscript] General Revision and comments (openjournals/joss-reviews#4822)
Recreation of #6 as issue on behalf of @govarguz

----

Hello everyone,

this a very creative framework for handling multiscale communication. We had also a similar multiscale challenge by sequentially going back and forth between AA-MD (also CG-MD) simulations and Poisson Boltzmann [HERE](https://github.com/pyF4all/Notebooks_phase1). However, we did not manage the automated communication included in GLUE, if I understand it correctly?

I will start now with general comments:

1. There is a conceptual difference between sequential and concurrent multiscale simulations, looking at figure 1 I would rather say GLUE seems to be sequential, at least until a 'specific model' has been trained, did I get this correctly?  Is there a concurrent capability of GLUE?  In any case please clarify shortly in page 1, section: Summary.
2. In the same lines, solving the load imbalance issue is indeed key for large scale simulations, nonetheless, another important point is the initial spatial decomposition of the simulation boxes, for this, it was shown by predictive domain decomposition that heterogeneous systems can be also initially optimized, which is not dynamic load balancing but still a good starting point, specially for concurrent multiscale is worth mentioning https://doi.org/10.1103/PhysRevE.96.053311
3. Does GLUE include a parser for different scale formats? If not, how is this handled?

This so far, thanks in advance for your point-to-point answers!

Part of https://github.com/openjournals/joss-reviews/issues/4822
