### 2022-12-06 lanl/GLUE#8 @govarguz
Hi @apachalieva, sure! this is a pretty clear case https://github.com/espressopp/espressopp you could use for the GLUECode_Library and also in general, as an example.
