### 2022-12-06 34deda9
Start of training readme
