+++ training/README.md
@@ -1,0 +1,4 @@
+# Training Data
+
+This directory contains what are effectively input decks to generate training data to pre-seed the active learning with.
+
