### 2022-12-06 3759e76
Added Logic for BGK training sets
