### 2022-12-06 93fc41e
Switched to math formatting
