### 2022-12-07 lanl/GLUE#11 @rspavel
#13 and associated issue improved description of how to use training data input decks and #16 clarifies how that sweep was selected.
