### 2022-12-07 lanl/GLUE#10 @junghans
I think we should stick with "interface" as the "I" in API stand for interface as well.
