### 2022-12-07 lanl/GLUE#10 @govarguz
I dislike "interface", because this is not specific. API is certainly more specific, or another less confusing definition.
