### 2022-12-07 lanl/GLUE#10 @rspavel
I guess I still disagree with that.

As I mentioned, basically every word of that title (except "GLUE") is ambiguous/"not specific"

Code: is this software or an encoding scheme
Framework: Is this a methodology? A set of tightly coupled services/systems?
Handling Communication: Is this an alternative to MPI? An interface to TCP? Etc
Interfaces: As discussed, this could mean a very scientific definition or it could be closer to how "interface" is used in API
Scales: As discussed, this could mean we are talking about physical scales or it could mean we are talking about different sized allocations/clusters.

Like most things, it only truly makes sense in the greater context. Because this is a "Code" that implements a "framework", that implies it is closer to software than encoding scheme. Communication, scale, and interface all combine to indicate we are using the more computer science oriented uses of the term.
Also, this being submitted to a Journal of Open Source Software further indicates we are using more of a software oriented definition of the word "interface"

And any confusion beyond that is rapidly negated by the contents of the abstract/summary, let alone the rest of the paper.

I am in the process of writing up some text to address #14 and we have already addressed every other issue. Unless we can think of a really good alternative title, I am going to push back against this requested change.
