### 2022-12-07 lanl/GLUE#10 @lubbersnick
I agree with @rspavel on all of these points. "Interface" and "Scale" in particular serve as useful "puns" that have both computer science and computational science meanings which are both applicable to this work. And also agree that the abstract/summary/paper can clarify. If we could reach the highest levels of linguistic precision in titles alone then abstracts would not be a thing at all. So for my vote I would also say I resist this change.
