### 2022-12-08 74e1c1b
Update paper.md
Added three subsections as part of the "Statement of need" discussing the following topics: "Concurrent multiscale simulations", "Load imbalance", "Multiple formats and parsing tools".
