+++ paper.md
@@ -217,11 +217,11 @@ Characteristics of the `GLUE` code:
 ## Concurrent multiscale simulations
 
 As depicted in Fig. <a href="#fig:glue_impl" data-reference-type="ref"
-data-reference="fig:glue_impl">1</a>, the parallelism of the Coarse Grain Solver is fully dependent on the solver of choice; however, the GLUECode_Library provides MPI interfaces to support distributed applications and all couplings thus far (barring hand-written tests for CI purposes) were with distributed coarse grain simulations. Initial efforts were made to make the GLUECode_Library calls non-blocking (e.g. mpi_irecv), but this was given low priority as the coarse grain solvers, considered so far, do not include extreme computations and can be executed in short times. This minimizes the need for overlapping computation and communication tasks using non-blocking functions (e.g. mpi_irecv). Similarly, there was work on providing a thread-safe interface for on-node parallelism (e.g. OpenMP), but this was similarly given low priority.
+data-reference="fig:glue_impl">1</a>, the parallelism of the Coarse Grain Solver is fully dependent on the solver of choice; however, the `GLUECode_Library` provides MPI interfaces to support distributed applications and all couplings thus far (barring hand-written tests for CI purposes) were with distributed coarse grain simulations. Initial efforts were made to make the `GLUECode_Library` calls non-blocking (similar to non-blocking communication in MPI), but this was given low priority as the coarse grain solvers considered thus far have comparatively fast time steps relative to the fine grain solvers and do not have a significant amount of work that can be overlapped with the cost of the `GLUE` Code. Similarly, there was work on providing a thread-safe interface for on-node parallelism (e.g. OpenMP), but this was similarly given low priority.
 
-A single GLUECode_Service is primarily a series of sequential operations that occur in a persistent service and is built around monitoring a task queue. Said task queue allows for parallelism via the use of HPC Job Schedulers, but the actual computation/generation of ML models, as well as processing of requests and results, is sequential. That said, multiple GLUECode_Service instances can be run in parallel, and this has been done as a way to lessen congestion for particularly large simulation.
+A single `GLUECode_Service` is primarily a series of sequential operations that occur in a persistent service and is built around monitoring a task queue. Said task queue allows for parallelism via the use of HPC Job Schedulers, but the actual computation/generation of ML models, as well as processing of requests and results, is sequential. That said, multiple `GLUECode_Service` instances can be run in parallel, and this has been done as a way to lessen congestion for particularly large simulation.
 
-The Fine Grain Solver is once again dependent on the solver. We have worked with both GPU-enabled LAMMPS and MPI-enabled LAMMPS and the SlurmScheduler and FluxScheduler parts of our json schema (https://github.com/lanl/GLUE/blob/manuscript/docs/inputSchema.json) are specifically set up to provide these configurations.
+The Fine Grain Solver is once again dependent on the solver. We have worked with both GPU-enabled LAMMPS and MPI-enabled LAMMPS and the SlurmScheduler and FluxScheduler parts of our json schema https://github.com/lanl/GLUE/blob/1.0/docs/inputSchema.json are specifically set up to provide these configurations.
 
 The overall GLUECode has a high degree of concurrency and resembles fork-join parallelism/MapReduce in practice, even if it consists of sequential tasks/stages.
 
@@ -229,12 +229,12 @@ The overall GLUECode has a high degree of concurrency and resembles fork-join pa
 
 We have implemented traditional approaches for load balancing used by asynchronous task-based runtimes that rely on some form of a task queue and then various work-stealing algorithms. However, due to the granularity of our tasks, we offload these tasks to HPC job schedulers like Slurm and Flux. By putting this work into a job queue, we can use proven tools to schedule those in an efficient manner.
 
[…]
