### 2022-12-08 lanl/GLUE#14 @apachalieva
We added information about concurrency, load imbalance, and formatting/parsing to the manuscript. With this, we close this issue.
