### 2022-12-08 #4822 @apachalieva
Dear @danielskatz, my colleagues and I believe we have addressed all reviewer requests. Here are the links to all the issues:

https://github.com/lanl/GLUE/issues/7
https://github.com/lanl/GLUE/issues/8
https://github.com/lanl/GLUE/issues/9
https://github.com/lanl/GLUE/issues/11
https://github.com/lanl/GLUE/issues/14

We appreciate the reviewers' input which improved the content and readability of the manuscript/repository.

There is a small dispute about the wording of the title - we disagree with one of the reviewers, and we would like to ask you to give us guidance on how to proceed on this issue. You can see the issue and the discussion here: https://github.com/lanl/GLUE/issues/10

Thank you,
GLUE Code development team
