### 2022-12-08 lanl/GLUE#10 @danielskatz
thanks to all for the interesting discussion here
