### 2022-12-09 lanl/GLUE#10 @apachalieva
We made small changes to the "Summary" to improve the readability and understanding of our code.

With this, I am closing this issue. Thank you for all the input and interesting discussion.
