### 2022-12-21 #4822 @junghans
> @apachalieva and @junghans - At this point could you:
>
> * [x]  Make a tagged release of your software, and list the version tag of the archived version here.
> * [x]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
> * [x]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
> * [x]  Please list the DOI of the archived version here.
>
> I can then move forward with recommending acceptance of the submission.

4 x Done

[![DOI](https://zenodo.org/badge/DOI/[10.5281/zenodo.7469110.svg)](https://doi.org/10.5281/zenodo.7469110](https://doi.org/10.5281/zenodo.7469110.svg)](https://doi.org/10.5281/zenodo.7469110))
