### 2022-12-23 #4822 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
