+++ paper.bib
@@ -47,7 +47,7 @@
 }
 
 @article{ross2017transition,
-  title={Transition from collisional to collisionless regimes in interpenetrating plasma flows on the national ignition facility},
+  title={Transition from collisional to collisionless regimes in interpenetrating plasma flows on the {N}ational {I}gnition {F}acility},
   author={Ross, JS and Higginson, DP and Ryutov, D and Fiuza, F and Hatarik, R and Huntington, CM and Kalantar, DH and Link, A and Pollock, BB and Remington, BA and others},
   journal={Physical Review Letters},
   volume={118},
@@ -71,9 +71,9 @@
 }
 
 @inproceedings{glosli2007extending,
-  title={Extending stability beyond {CPU} millennium:{A} micron-scale atomistic simulation of {K}elvin-{H}elmholtz instability},
+  title={Extending stability beyond {CPU} millennium: {A} micron-scale atomistic simulation of {K}elvin-{H}elmholtz instability},
   author={Glosli, James N and Richards, David F and Caspersen, Kyle J and Rudd, Robert E and Gunnels, John A and Streitz, Frederick H},
-  booktitle={SC'07: Proceedings of the 2007 ACM/IEEE conference on Supercomputing},
+  booktitle={SC'07: Proceedings of the 2007 ACM/IEEE Conference on Supercomputing},
   pages={1--11},
   year={2007},
   doi={10.1145/1362622.1362700},
@@ -156,9 +156,10 @@
   doi={10.1145/2834899.2834908}
 }
 
-@article{pavel2017using,
-  title={Using charm++ to support multiscale multiphysics},
+@inproceeding{pavel2017using,
+  title={Using {C}harm++ to support multiscale multiphysics on the {T}rinity supercomputer},
   author={Pavel, Robert and Junghans, Christoph and Mniszewski, Susan M and Germann, Timothy C},
+  booktitle={Proceedings of the Programming Models and Co-Design Meeting}
   year={2017}
 }
 
