### 2022-12-23 #4822 @danielskatz
@junghans - After proofreading, I've suggested some minor changes in https://github.com/lanl/GLUE/pull/20 - please merge this, or let me know what you disagree with, and then we can proceed.
