+++ paper.bib
@@ -156,7 +156,7 @@
   doi={10.1145/2834899.2834908}
 }
 
-@inproceeding{pavel2017using,
+@inproceedings{pavel2017using,
   title={Using {C}harm++ to support multiscale multiphysics on the {T}rinity supercomputer},
   author={Pavel, Robert and Junghans, Christoph and Mniszewski, Susan M and Germann, Timothy C},
   booktitle={Proceedings of the Programming Models and Co-Design Meeting},
